refactor indent into it's own component
