+++ public/app/core/components/AppChrome/DockedMegaMenu/MegaMenuItem.tsx
@@ -5,6 +5,8 @@ import { useLocalStorage } from 'react-use';
 import { GrafanaTheme2, NavModelItem } from '@grafana/data';
 import { Button, Icon, useStyles2, Text } from '@grafana/ui';
 
+import { Indent } from '../../Indent/Indent';
+
 import { FeatureHighlight } from './FeatureHighlight';
 import { MegaMenuItemIcon } from './MegaMenuItemIcon';
 import { MegaMenuItemText } from './MegaMenuItemText';
@@ -47,6 +49,7 @@ export function MegaMenuItem({ link, activeItem, level = 0, onClose }: Props) {
             <FeatureHighlightWrapper>
               <div className={styles.iconWrapper}>{level === 0 && <MegaMenuItemIcon link={link} />}</div>
             </FeatureHighlightWrapper>
+            <Indent level={Math.max(0, level - 1)} spacing={2} />
             <Text truncate>{link.text}</Text>
           </div>
         </MegaMenuItemText>
